+++ src/extension.ts
@@ -25,6 +25,7 @@ import { correctTypos } from './correctTypos';
 import { copyLog } from './copyLog';
 import { convertLog } from './convertLog';
 import { editAudioBox } from './editAudioBox';
+import { editContent } from './editContent';
 
 /**
  * 从某个HTML文件读取能被Webview加载的HTML内容
@@ -80,10 +81,11 @@ export function activate(context: vscode.ExtensionContext) {
 
 	//正则替换
 	context.subscriptions.push(vscode.commands.registerCommand('trpg-replay-generator-log.replaceAngleBrackets',regexReplace.replaceAngleBrackets));
-	context.subscriptions.push(vscode.commands.registerCommand('trpg-replay-generator-log.addAsteriskMarks',regexReplace.addAsteriskMarks));
-	context.subscriptions.push(vscode.commands.registerCommand('trpg-replay-generator-log.addSoundEffectsInBatches',regexReplace.addSoundEffectsInBatches));
 	context.subscriptions.push(vscode.commands.registerCommand('trpg-replay-generator-log.replaceDiceMaidLine',regexReplace.replaceDiceMaidLine));
+	//编辑音效框
 	context.subscriptions.push(vscode.commands.registerCommand('trpg-replay-generator-log.editAudioBox',editAudioBox));
+	//编辑对话行文本内容
+	context.subscriptions.push(vscode.commands.registerCommand('trpg-replay-generator-log.editContent',editContent));
 	//转换Log格式
 	context.subscriptions.push(vscode.commands.registerCommand('trpg-replay-generator-log.convertLog',convertLog));